+++ tag-20180828/alibabacloud_tag20180828/models.py
@@ -591,7 +591,7 @@ class CreateTagsRequestTagKeyValueParamListTagValueParamList(TeaModel):
         self.description = description
         # The value of tag N.
         # 
-        # The tag value can be up to 128 characters in length and cannot contain `http://` or `https://`. 
+        # The tag value can be up to 128 characters in length and cannot contain `http://` or `https://`.
         # 
         # Valid values of N: 1 to 10.
         self.value = value
@@ -631,15 +631,15 @@ class CreateTagsRequestTagKeyValueParamList(TeaModel):
         # 
         # Valid values of N: 1 to 10.
         self.description = description
-        # The key of tag N.
+        # The value of tag N.
         # 
-        # The tag key can be up to 128 characters in length and cannot contain `http://` or `https://`. The tag key cannot start with `acs:` or `aliyun`.
+        # The tag value can be up to 128 characters in length and cannot start with `acs:` or `aliyun`. The tag key cannot contain `http://` or `https://`.
         # 
         # Valid values of N: 1 to 10.
         # 
         # This parameter is required.
         self.key = key
-        # The information about the tag value.
+        # The information about the tag values.
         self.tag_value_param_list = tag_value_param_list
 
     def validate(self):
@@ -2101,6 +2101,7 @@ class GetEffectivePolicyRequest(TeaModel):
         owner_id: int = None,
         region_id: str = None,
         resource_owner_account: str = None,
+        tag_keys: List[str] = None,
         target_id: str = None,
         target_type: str = None,
     ):
@@ -2109,6 +2110,7 @@ class GetEffectivePolicyRequest(TeaModel):
         # The region ID. Set the value to cn-shanghai.
         self.region_id = region_id
         self.resource_owner_account = resource_owner_account
+        self.tag_keys = tag_keys
         # The ID of the object.
         # 
         # >  If you use the Tag Policy feature in single-account mode, this parameter is optional. If you use the Tag Policy feature in multi-account mode, this feature is required.
@@ -2140,6 +2142,8 @@ class GetEffectivePolicyRequest(TeaModel):
             result['RegionId'] = self.region_id
         if self.resource_owner_account is not None:
             result['ResourceOwnerAccount'] = self.resource_owner_account
+        if self.tag_keys is not None:
+            result['TagKeys'] = self.tag_keys
         if self.target_id is not None:
             result['TargetId'] = self.target_id
         if self.target_type is not None:
@@ -2156,6 +2160,8 @@ class GetEffectivePolicyRequest(TeaModel):
             self.region_id = m.get('RegionId')
         if m.get('ResourceOwnerAccount') is not None:
             self.resource_owner_account = m.get('ResourceOwnerAccount')
+        if m.get('TagKeys') is not None:
+            self.tag_keys = m.get('TagKeys')
         if m.get('TargetId') is not None:
             self.target_id = m.get('TargetId')
         if m.get('TargetType') is not None:
@@ -2163,16 +2169,22 @@ class GetEffectivePolicyRequest(TeaModel):
         return self
 
 
-class GetEffectivePolicyResponseBody(TeaModel):
+class GetEffectivePolicyResponseBodyPolicyAttachmentsPolicyList(TeaModel):
     def __init__(
         self,
-        effective_policy: str = None,
-        request_id: str = None,
+        attach_seq: int = None,
+        attach_time: str = None,
+        policy_id: str = None,
+        policy_name: str = None,
+        target_id: str = None,
+        target_type: str = None,
     ):
-        # The effective tag policy.
-        self.effective_policy = effective_policy
-        # The ID of the request.
-        self.request_id = request_id
+        self.attach_seq = attach_seq
+        self.attach_time = attach_time
+        self.policy_id = policy_id
+        self.policy_name = policy_name
+        self.target_id = target_id
+        self.target_type = target_type
 
     def validate(self):
         pass
@@ -2183,8 +2195,115 @@ class GetEffectivePolicyResponseBody(TeaModel):
             return _map
 
         result = dict()
+        if self.attach_seq is not None:
+            result['AttachSeq'] = self.attach_seq
+        if self.attach_time is not None:
+            result['AttachTime'] = self.attach_time
+        if self.policy_id is not None:
+            result['PolicyId'] = self.policy_id
+        if self.policy_name is not None:
+            result['PolicyName'] = self.policy_name
+        if self.target_id is not None:
+            result['TargetId'] = self.target_id
+        if self.target_type is not None:
+            result['TargetType'] = self.target_type
+        return result
+
+    def from_map(self, m: dict = None):
+        m = m or dict()
+        if m.get('AttachSeq') is not None:
+            self.attach_seq = m.get('AttachSeq')
+        if m.get('AttachTime') is not None:
+            self.attach_time = m.get('AttachTime')
+        if m.get('PolicyId') is not None:
+            self.policy_id = m.get('PolicyId')
+        if m.get('PolicyName') is not None:
+            self.policy_name = m.get('PolicyName')
+        if m.get('TargetId') is not None:
+            self.target_id = m.get('TargetId')
+        if m.get('TargetType') is not None:
+            self.target_type = m.get('TargetType')
+        return self
+
+
+class GetEffectivePolicyResponseBodyPolicyAttachments(TeaModel):
+    def __init__(
+        self,
+        policy_list: List[GetEffectivePolicyResponseBodyPolicyAttachmentsPolicyList] = None,
+        policy_type: str = None,
+        tag_key: str = None,
+    ):
+        self.policy_list = policy_list
+        self.policy_type = policy_type
+        self.tag_key = tag_key
+
+    def validate(self):
+        if self.policy_list:
+            for k in self.policy_list:
+                if k:
+                    k.validate()
+
+    def to_map(self):
+        _map = super().to_map()
+        if _map is not None:
+            return _map
+
+        result = dict()
+        result['PolicyList'] = []
+        if self.policy_list is not None:
+            for k in self.policy_list:
+                result['PolicyList'].append(k.to_map() if k else None)
+        if self.policy_type is not None:
+            result['PolicyType'] = self.policy_type
+        if self.tag_key is not None:
+            result['TagKey'] = self.tag_key
+        return result
+
+    def from_map(self, m: dict = None):
+        m = m or dict()
+        self.policy_list = []
+        if m.get('PolicyList') is not None:
+            for k in m.get('PolicyList'):
+                temp_model = GetEffectivePolicyResponseBodyPolicyAttachmentsPolicyList()
+                self.policy_list.append(temp_model.from_map(k))
+        if m.get('PolicyType') is not None:
+            self.policy_type = m.get('PolicyType')
+        if m.get('TagKey') is not None:
+            self.tag_key = m.get('TagKey')
+        return self
+
+
+class GetEffectivePolicyResponseBody(TeaModel):
+    def __init__(
+        self,
+        effective_policy: str = None,
+        policy_attachments: List[GetEffectivePolicyResponseBodyPolicyAttachments] = None,
+        request_id: str = None,
+    ):
+        # The effective tag policy.
+        self.effective_policy = effective_policy
+        self.policy_attachments = policy_attachments
+        # The ID of the request.
+        self.request_id = request_id
+
+    def validate(self):
+        if self.policy_attachments:
+            for k in self.policy_attachments:
+                if k:
+                    k.validate()
+
+    def to_map(self):
+        _map = super().to_map()
+        if _map is not None:
+            return _map
+
+        result = dict()
         if self.effective_policy is not None:
             result['EffectivePolicy'] = self.effective_policy
+        result['PolicyAttachments'] = []
+        if self.policy_attachments is not None:
+            for k in self.policy_attachments:
+                result['PolicyAttachments'].append(k.to_map() if k else None)
         if self.request_id is not None:
             result['RequestId'] = self.request_id
         return result
@@ -2193,6 +2312,11 @@ class GetEffectivePolicyResponseBody(TeaModel):
         m = m or dict()
         if m.get('EffectivePolicy') is not None:
             self.effective_policy = m.get('EffectivePolicy')
+        self.policy_attachments = []
+        if m.get('PolicyAttachments') is not None:
+            for k in m.get('PolicyAttachments'):
+                temp_model = GetEffectivePolicyResponseBodyPolicyAttachments()
+                self.policy_attachments.append(temp_model.from_map(k))
         if m.get('RequestId') is not None:
             self.request_id = m.get('RequestId')
         return self
@@ -3428,7 +3552,7 @@ class ListResourcesByTagRequestTagFilter(TeaModel):
         self.key = key
         # The tag value. This parameter specifies a filter condition for the query.
         # 
-        # The tag value can be a maximum of 128 characters in length. It cannot contain `http://` or `https://`.
+        # The tag value can be up to 128 characters in length and cannot contain `http://` or `https://`.
         self.value = value
 
     def validate(self):
@@ -3901,6 +4025,7 @@ class ListSupportResourceTypesResponseBodySupportResourceTypes(TeaModel):
         resource_type: str = None,
         support_items: List[ListSupportResourceTypesResponseBodySupportResourceTypesSupportItems] = None,
     ):
+        # The resource ARN template.
         self.arn_template = arn_template
         # The service code.
         self.product_code = product_code
@@ -4050,11 +4175,9 @@ class ListTagKeysRequestTagFilter(TeaModel):
         self,
         key: str = None,
     ):
-        # The tag key.
+        # The tag key for a fuzzy query.
         # 
         # This parameter is used together with the `FuzzyType` parameter.
-        # 
-        # >  This parameter is available only in the China (Shenzhen) and China (Hong Kong) regions.
         self.key = key
 
     def validate(self):
@@ -4103,10 +4226,8 @@ class ListTagKeysRequest(TeaModel):
         self.category = category
         # The type of the query. Valid values:
         # 
-        # *   EQUAL: exact match. This is the default value.
-        # *   PREFIX: prefix-based fuzzy match.
-        # 
-        # >  This parameter is available only in the China (Shenzhen) and China (Hong Kong) regions.
+        # *   EQUAL (default): exact match
+        # *   PREFIX: prefix-based fuzzy match
         self.fuzzy_type = fuzzy_type
         # The token that is used to start the next query.
         self.next_token = next_token
@@ -4658,11 +4779,9 @@ class ListTagValuesRequestTagFilter(TeaModel):
         self,
         value: str = None,
     ):
-        # The tag value.
+        # The tag value for a fuzzy query.
         # 
         # This parameter is used together with the `FuzzyType` parameter.
-        # 
-        # >  This parameter is available only in the China (Shenzhen) and China (Hong Kong) regions.
         self.value = value
 
     def validate(self):
@@ -4703,10 +4822,8 @@ class ListTagValuesRequest(TeaModel):
         self.tag_filter = tag_filter
         # The type of the query. Valid values:
         # 
-        # *   EQUAL: exact match. This is the default value.
-        # *   PREFIX: prefix-based fuzzy match.
-        # 
-        # >  This parameter is available only in the China (Shenzhen) and China (Hong Kong) regions.
+        # *   EQUAL (default): exact match
+        # *   PREFIX: prefix-based fuzzy match
         self.fuzzy_type = fuzzy_type
         # The tag key. This parameter specifies a filter condition for the query.
         # 
@@ -5720,7 +5837,9 @@ class UntagResourcesRequest(TeaModel):
         # This parameter is required.
         self.resource_arn = resource_arn
         self.resource_owner_account = resource_owner_account
-        # A tag key.
+        # The key of tag N.
+        # 
+        # Valid values of N: 1 to 10.
         # 
         # This parameter is required.
         self.tag_key = tag_key